--- src/types/committee/sign-up/index.ts
@@ -1,7 +1,0 @@
-export interface CommitteeSignUpFormData {
-  name: string;
-  email: string;
-  password: string;
-  departmentId: number;
-  phoneNumber: string;
-}
