+++ src/types/common/sign-up/index.tsx
@@ -1,0 +1,7 @@
+export interface SignUpFormData {
+  name: string;
+  email: string;
+  password: string;
+  departmentId: number;
+  phoneNumber: string;
+}
